+++ tests/test_vis_cpu.py
@@ -6,6 +6,7 @@ from astropy.units import sday
 from pyuvsim.analyticbeam import AnalyticBeam
 
 from vis_cpu import conversions, simulate_vis, vis_cpu
+from vis_cpu.vis_cpu import construct_pixel_beam_spline
 
 np.random.seed(0)
 NTIMES = 10
@@ -87,6 +88,20 @@ def test_vis_cpu():
             polarized=True,
         )
 
+    # Check that a complex pixel beam works
+    for i in range(freq.size):
+        _vis = vis_cpu(
+            antpos,
+            freq[i],
+            eq2tops,
+            crd_eq,
+            I_sky[:, i],
+            bm_cube=(1.0 * 0.1j) * beam_cube[:, i, :, :],
+            precision=1,
+            polarized=False,
+        )
+    assert np.all(~np.isnan(_vis))  # check that there are no NaN values
+
 
 def test_simulate_vis():
     """Test basic operation of simple wrapper around vis_cpu, `simulate_vis`."""
@@ -139,3 +154,31 @@ def test_simulate_vis():
         latitude=-30.7215 * np.pi / 180.0,
     )
     assert np.all(~np.isnan(vis))  # check that there are no NaN values
+
+
+def test_construct_pixel_beam_spline():
+    """Test pixel beam interpolation."""
+    freqs = np.linspace(100.0e6, 120.0e6, NFREQ)  # Hz
+
+    # Create pixel beam
+    beam = AnalyticBeam(type="gaussian", diameter=14.0)
+    beams = [beam, beam, beam]  # 3 ants
+
+    beam_pix = [
+        conversions.uvbeam_to_lm(bm, freqs, n_pix_lm=20, polarized=False)
+        for bm in beams
+    ]
+    beam_cube = np.array(beam_pix)
+
+    # Complex values should raise an error
+    with pytest.raises(TypeError):
+        construct_pixel_beam_spline(beam_cube[:, 0] + 1e-5j)  # only 1 freq
+
+    # Mock-up a polarized beam
+    bm_cube_pol = beam_cube[np.newaxis, np.newaxis, :, 0, :, :]
+    beam_splines_pol = construct_pixel_beam_spline(bm_cube_pol)
+
+    # Check for expected no. of dimensions
+    assert len(beam_splines_pol) == 1
+    assert len(beam_splines_pol[0]) == 1
+    assert len(beam_splines_pol[0][0]) == len(beams)